requirements.txt
torch~=2.2.2
torchvision~=0.17.2
matplotlib~=3.9.2
numpy~=1.26.4
tqdm~=4.67.1
pillow~=10.4.0
